+++ frontend/src/pages/ErrorPage/components/ErrorPageContainer/ErrorPageContainer.styles.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const StyledDescription = styled.p`
+  text-align: center;
+  font-size: 2rem;
+  line-height: 2.8rem;
+`;
+
+export { StyledDescription };
